Massive refactor to the https://github.com/golang-standards/project-layout standard layout
