+++ internal/app/coinjar/coinjar_suite_test.go
