+++ internal/app/coinjar/transaction.go
